backend/chat/requirements.txt
Django==5.1.4
daphne==4.1.2
channels==4.2.0
PyJWT==2.10.1
Twisted==24.11.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
python-dotenv==1.0.1
